+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## TBD
+## 4.9.0 (2018-10-29)
 
 ### Enhancements